mailchatbot/requirements.txt
google-generativeai
langchain
langchain-google-genai
langchain-community
faiss-cpu
python-dotenv
beautifulsoup4
